Refactor pharmacy to add new feature
