+++ index.js
@@ -1,11 +1,12 @@
-import { Drug, Pharmacy } from "./pharmacy";
+import { HerbalTea, MagicPill, Fervex, Drug } from "./drugs";
+import { Pharmacy } from "./pharmacy";
 import fs from "fs";
 
 const drugs = [
   new Drug("Doliprane", 20, 30),
-  new Drug("Herbal Tea", 10, 5),
-  new Drug("Fervex", 5, 40),
-  new Drug("Magic Pill", 15, 40)
+  new HerbalTea("Herbal Tea", 10, 5),
+  new Fervex("Fervex", 5, 40),
+  new MagicPill("Magic Pill", 15, 40)
 ];
 const trial = new Pharmacy(drugs);
